Perbaiki Confirm Delete Data Inspeksi

+++ application/views/admin/v_data_inspeksi.php
@@ -203,7 +203,7 @@
                       </td>
                       <td class="text-center"><?= $row->periode ?></td>
                       <td class="text-center">
-                        <a href="<?= base_url(); ?>admin/hapus_inspeksi/<?= $row->id ?>" class="btn btn-sm btn-danger" data-tooltip="Hapus Data Inspeksi"><i class="fa fa-trash"></i></a>
+                        <a href="<?= base_url(); ?>admin/hapus_inspeksi/<?= $row->id ?>" class="btn btn-sm btn-danger" data-tooltip="Hapus Data Inspeksi" onclick="return confirm('Apakah anda yakin hapus data inspeksi ini?')"><i class="fa fa-trash"></i></a>
                       </td>
                     </tr>
                   <?php endforeach; ?>
